+++ src/Auth0.php
@@ -13,8 +13,10 @@ namespace angellco\auth0;
 use angellco\auth0\models\Settings;
 use angellco\auth0\services\Auth as AuthService;
 
+use angellco\auth0\variables\Auth0Variable;
 use Craft;
 use craft\base\Plugin;
+use craft\web\twig\variables\CraftVariable;
 use yii\base\Event;
 use yii\web\User;
 use yii\web\UserEvent;
@@ -56,6 +58,17 @@ class Auth0 extends Plugin
     {
         parent::init();
         self::$plugin = $this;
+
+        // Register the variable
+        Event::on(
+            CraftVariable::class,
+            CraftVariable::EVENT_INIT,
+            function (Event $event) {
+                /** @var CraftVariable $variable */
+                $variable = $event->sender;
+                $variable->set('auth0', new Auth0Variable());
+            }
+        );
 
         // Bind to the after logout event so we can clear the Auth0 session
         Event::on(