+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 
 ## 1.0.2 - 2020-06-05
+### Added
+- Added a variable and service method that attempts to silently log in to Craft if there is already an active Auth0 session: `{% do craft.auth0.silentLogin() %}`
+
 ### Changed
 - When logging out from the core Craft session there is now a check for an active Auth0 session before attempting to logout from there too